🚚 rename: 컨벤션 준수 함수명 및 파일명 수정
